lib/pyproject.toml
[project]
name = "lib"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Yang Qin", email = "qin.yang@emotion-tech.co.jp" }
]
dependencies = [
    "openai>=1.11.1",
    "pydantic>=2.6.1",
    "ulid-py>=1.1.0",
]
readme = "README.md"
requires-python = " >= 3.11"

[project.scripts]
hello = "lib:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/lib"]
